address filter fix
